app/ventas/requirements.txt
psycopg2-binary
prometheus-client
